Improve handling of client messages
