+++ app/player/main/main-window-events.js
@@ -3,8 +3,6 @@
 
   var contentEventHandlers = [];
 
-  contentEventHandlers.push(require("../platform/content-event-handlers/bypass-cors.js"));
-
   window.addEventListener("load", function() {
     showOptionsMenu();
   });
@@ -26,26 +24,4 @@
       outerBounds: boundsSpecification
     });
   }
-
-  window.addEventListener("message", function(evt) {
-    var handlers = contentEventHandlers.filter(function(handler) {
-      return handler.handles(evt);
-    });
-
-    if(handlers.length === 0) {
-      respondWithError("No handlers were found for the event");
-      return false;
-    }
-    else if(handlers.length > 1) {
-      respondWithError("Only one handler can exist for the given event");
-      return false;
-    }
-
-    return handler[0].process(evt);
-
-    function respondWithError(err) {
-      evt.data.error = err;
-      evt.source.postMessage(evt.data, "*");
-    }
-  });
 }());
