+++ app/player/platform/ui-controller.js
@@ -1,3 +1,8 @@
+var contentEventHandlers = [];
+
+contentEventHandlers.push(require("../platform/content-event-handlers/bypass-cors.js"));
+contentEventHandlers.push(require("../platform/content-event-handlers/storage-component-response.js"));
+
 module.exports = {
   createViewWindow: function(contentTarget) {
     if (!contentTarget) {
@@ -16,9 +21,33 @@ module.exports = {
     document.body.appendChild(view);
 
     view.addEventListener("loadstop", sendRegistrationMessage);
+    view.addEventListener("message", function(evt) {
+      console.log("Received event", evt);
+
+      var handlers = contentEventHandlers.filter(function(handler) {
+        return handler.handles(evt);
+      });
+
+      if(handlers.length === 0) {
+        respondWithError("No handlers were found for the event");
+        return false;
+      }
+      else if(handlers.length > 1) {
+        respondWithError("Only one handler can exist for the given event");
+        return false;
+      }
+
+      return handler[0].process(evt);
+
+      function respondWithError(err) {
+        evt.data.error = err;
+        evt.source.postMessage(evt.data, "*");
+      }
+    });
 
     function sendRegistrationMessage() {
       view.contentWindow.postMessage("register.chrome.app.window", "*");
+      view.contentWindow.postMessage({ type: "offline-player-init" }, "*");
       view.removeEventListener("loadstop", sendRegistrationMessage);
     }
